ui test moved under another folder in file structure
